# 🎮 テトリスゲーム機能修正 + PerformanceOptimizer本格稼働版