… demo/production-tetris-app.html
… docs/DESIGN_INDEX.md
… docs/PERFORMANCE_OPTIMIZER_FINAL_REPORT.md
… docs/PRODUCTION_DEPLOYMENT_COMPLETE.md
… docs/context.md
… docs/plan.md
… reports/FINAL_VERIFICATION_REPORT.md
… reports/performance-report-2025-09-03.html
… reports/performance-report-2025-09-03.json
… reports/performance-report-2025-09-03.md
… reports/ux-evaluation-report.md
… src/presentation/main.js
… tests/integration/performance-optimizer-integration.test.js
… tests/performance/real-game-performance-verification.test.js
… tools/generate-performance-report.js
… tools/performance-verification.html
… tools/production-monitoring-dashboard.html